// src/app/Models/logoes/logo.ts
export class Logo {
    IdLogo: number;
    NameLogo: string;
    Description: string;
    ImageURL: string;
    ArchiveURL: string;
    IdCompany: number;
    IdValueLogo: number;
    IdGarmentColor: number;
    Status: boolean;
    UserCreate: string;
    CreateDate: Date;
    UserModified: string;
    ModifiedDate: Date;
}